Moved DomainAttributes to Authorization Api module
